Add Handbook item; refactor inventory & UI
Adds a new Handbook item (prefab, meta and asset) and registers it in NetworkPrefabs; assigns the handbookPrefab reference on the Engineer prefab. Refactors InventoryManager: adds starting-item handling (spawns handbook into inventory), introduces slot-based EquipSlot/Refresh/Hide flows, improves pickup/placement logic (uses target slot selection, applies ItemLoot offsets), and consolidates visibility handling (SetCurrentItemVisibility). InventorySlot UI now uses DOTween scaling for highlight effects and drops the highlight GameObject. PlayerSpawningState now spawns players by role via SpawnByRole. UI prefab/MainScene prefab instance updated to remove old highlight objects and set relevant CanvasGroup alpha to 0. Also tweaks Sonar material _ScanAngle value.

+++ SUBMARİNE PUZZLE GAME/Assets/Scripts/UtiltyEvents/StressManager.cs
@@ -146,7 +146,7 @@ public class StressManager : NetworkBehaviour
     {
         if (availableEvents.Count == 0) return;
 
-        float roll = Random.Range(0f, 100f);
+        int roll = Random.Range(0, 100);
         Debug.Log($"[ZAR] Atılan: {roll}, Stres: {currentStress.value}");
 
         if (roll < currentStress.value)
